+++ spring-web/src/test/java/org/springframework/http/server/reactive/ServerHttpRequestIntegrationTests.java
@@ -56,7 +56,7 @@ public class ServerHttpRequestIntegrationTests extends AbstractHttpHandlerIntegr
 			URI uri = request.getURI();
 			assertThat(uri.getScheme()).isEqualTo("http");
 			assertNotNull(uri.getHost());
-			assertThat((long) uri.getPort()).isNotEqualTo((long) -1);
+			assertThat(uri.getPort()).isNotEqualTo((long) -1);
 			assertNotNull(request.getRemoteAddress());
 			assertThat(uri.getPath()).isEqualTo("/foo");
 			assertThat(uri.getQuery()).isEqualTo("param=bar");
